fix: WebXR examples (#218)
* fix ARButton constructor parameters

ARButton can specify WebXR session initialization options when the
button is clicked. In particular, it offers the ability to define the
`domOverlay` property allow DOM elements to render on top of the AR
session.

* Add missing `envMap` field to `XRControllerModelFactory`

* Provide more-specific type definition for `motionController` field

In OculusHandModel and OculusHandPointerModel, the motionController field
will be set to an XRHandMeshModel value.

* Specify type for `XREstimatedLight`'s `lightProbe` field.

The lightProbe field is a regular Three.js LightProbe object.

* Fix `XRHandModelFactory`'s `createHandModel`'s `profile` parameter

The value of "oculus" has not been used for nearly a year. A more
generic "mesh" value has been available since r129.

* add to contributors

* Use the complete WebXR types from DefinitelyTyped

* revert yarn.lock

* try running yarn install